data generation

+++ pipelines/prepare-data/prepare-musdb.sh
@@ -2,8 +2,20 @@
 # [wf] execute setup stage
 
 mkdir -p raw
-mkdir -p raw/musdb
-#unzip raw/musdb18.zip
-tar xvf raw/musdb18.zip -C raw/musdb
-docker pull faroit/sigsep-mus-io
-docker run --rm -v `pwd`/raw/musdb:/data faroit/sigsep-mus-io /scripts/decode.sh
+if [ ! -d raw/musdb ]; then
+    mkdir -p raw/musdb
+    unzip raw/musdb18.zip -d raw/musdb/
+    docker pull faroit/sigsep-mus-io
+    docker run --rm -v `pwd`/raw/musdb:/data faroit/sigsep-mus-io /scripts/decode.sh
+fi
+
+mkdir -p data
+if [ ! -d data/musdb ]; then
+    docker run --workdir=/pipeline -v `pwd`:/pipeline \
+    -e USERID=$UID \
+  thesis scripts/organize_musdb.py
+fi
+
+if [ ! -d ../../data ]; then
+    ln -s `pwd`/data `pwd`/../..
+fi
